POST /room_types request done

+++ src/main/java/ru/cft/shift/luxury_dorm/entity/UserEntity.java
@@ -14,7 +14,7 @@ public class UserEntity {
     @Column(name = "Name")
     private String name;
     @Column(name = "Balance")
-    private float balance;
+    private Float balance;
     @OneToOne(cascade = CascadeType.ALL)
     private RoomEntity room;
 
@@ -34,11 +34,11 @@ public class UserEntity {
         this.name = name;
     }
 
-    public float getBalance() {
+    public Float getBalance() {
         return balance;
     }
 
-    public void setBalance(Integer balance) {
+    public void setBalance(Float balance) {
         this.balance = balance;
     }
